@@ -6,7 +6,7 @@ import 'package:ctpaga/views/menu/menu.dart';
 import 'package:ctpaga/providers/provider.dart';
 import 'package:ctpaga/env.dart';
 
-import 'package:flutter_masked_text/flutter_masked_text.dart';
+import 'package:flutter_masked_text2/flutter_masked_text2.dart';
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter/material.dart';
@@ -17,7 +17,7 @@ class ProcessSalesPage extends StatefulWidget {
 }
 
 class _ProcessSalesPageState extends State<ProcessSalesPage> {
-  double _total;
+  double? _total;
   bool _statusButton = false, _statusButtonRegister = false;
 
   @override
