@@ -1,7 +1,7 @@
 class Picture{
   // ignore: non_constant_identifier_names
-  int id, commerce_id;
-  String description, url;
+  int? id, commerce_id;
+  String? description, url;
 
   // ignore: non_constant_identifier_names
   Picture({this.id, this.description, this.url, this.commerce_id});
