@@ -4,7 +4,7 @@ import 'package:ctpaga/views/navbar/navbar.dart';
 import 'package:ctpaga/providers/provider.dart';
 import 'package:ctpaga/env.dart';
 
-import 'package:flutter_masked_text/flutter_masked_text.dart';
+import 'package:flutter_masked_text2/flutter_masked_text2.dart';
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:provider/provider.dart';
 import 'package:http/http.dart' as http;
@@ -321,8 +321,8 @@ class _NewShippingPageState extends State<NewShippingPage> {
     setState(() => _statusButtonSave = true);
     await Future.delayed(Duration(milliseconds: 150));
     setState(() => _statusButtonSave = false);
-    if (_formKeyShipping.currentState.validate()) {
-      _formKeyShipping.currentState.save();
+    if (_formKeyShipping.currentState!.validate()) {
+      _formKeyShipping.currentState!.save();
       
       if(_switchFree)
         _price = "FREE";
