@@ -86,7 +86,7 @@ class _RegisterPageState extends State<RegisterPage> {
                   borderSide: BorderSide(color: colorLogo),
                 ),
               ),
-              onSaved: (String value) => _name = value,
+              onSaved: (String? value) => _name = value,
               validator: _validateName,
               textInputAction: TextInputAction.next,
               cursorColor: colorLogo,
@@ -115,7 +115,7 @@ class _RegisterPageState extends State<RegisterPage> {
                   borderSide: BorderSide(color: colorLogo),
                 ),
               ),
-              onSaved: (String value) => _address = value,
+              onSaved: (String? value) => _address = value,
               validator: _validateAddress,
               textInputAction: TextInputAction.next,
               cursorColor: colorLogo,
@@ -144,7 +144,7 @@ class _RegisterPageState extends State<RegisterPage> {
                   borderSide: BorderSide(color: colorLogo),
                 ),
               ),
-              onSaved: (String value) => _phone = value,
+              onSaved: (String? value) => _phone = value,
               validator: _validatePhone,
               textInputAction: TextInputAction.next,
               cursorColor: colorLogo,
@@ -177,7 +177,7 @@ class _RegisterPageState extends State<RegisterPage> {
                   ),
               ),
               validator: _validateEmail,
-              onSaved: (String value) => _email = value.toLowerCase().trim(),
+              onSaved: (String? value) => _email = value.toLowerCase().trim(),
               textInputAction: TextInputAction.next,
               cursorColor: colorLogo,
               style: TextStyle(
@@ -223,7 +223,7 @@ class _RegisterPageState extends State<RegisterPage> {
                   ),
               ),
               validator: _validatePassword,
-              onSaved: (String value) => _password = value,
+              onSaved: (String? value) => _password = value,
               textInputAction: TextInputAction.next,
               cursorColor: colorLogo,
               style: TextStyle(
@@ -268,7 +268,7 @@ class _RegisterPageState extends State<RegisterPage> {
                   ),
                 ),
               validator: _validatePasswordConfirm,
-              onSaved: (String value) => _passwordConfirm = value,
+              onSaved: (String? value) => _passwordConfirm = value,
               textInputAction: TextInputAction.done,
               onFieldSubmitted: (term){
                 FocusScope.of(context).requestFocus(new FocusNode()); //save the keyboard
@@ -393,8 +393,8 @@ class _RegisterPageState extends State<RegisterPage> {
     setState(() {
       _statusError = false;
     });
-    if (_formKeyRegister.currentState.validate()) {
-      _formKeyRegister.currentState.save();
+    if (_formKeyRegister.currentState!.validate()) {
+      _formKeyRegister.currentState!.save();
       
       _onLoading();
 
@@ -511,7 +511,7 @@ class _RegisterPageState extends State<RegisterPage> {
     );
   }
 
-  String _validateName(String value) {
+  String _validateName(String? value) {
     // This is just a regular expression for name
     String p = '[a-zA-Z]';
     RegExp regExp = new RegExp(p);
@@ -525,7 +525,7 @@ class _RegisterPageState extends State<RegisterPage> {
     return 'Ingrese nombre y apellido válido';
   }
 
-  String _validateAddress(String value) {
+  String _validateAddress(String? value) {
 
     if (value.length >=3) {
       // So, the address is valid
@@ -536,7 +536,7 @@ class _RegisterPageState extends State<RegisterPage> {
     return 'Ingrese una dirección válido';
   }
 
-  String _validatePhone(String value) {
+  String _validatePhone(String? value) {
     // This is just a regular expression for phone
     //String p = r'^(?:(\+)58|0)(?:2(?:12|4[0-9]|5[1-9]|6[0-9]|7[0-8]|8[1-35-8]|9[1-5]|3[45789])|4(?:1[246]|2[46]))\d{7}$';
     String p = r'^(0414|0424|0412|0416|0426)[0-9]{7}$';
@@ -551,7 +551,7 @@ class _RegisterPageState extends State<RegisterPage> {
     return 'Ingrese un número de teléfono válido';
   }
 
-  String _validateEmail(String value) {
+  String _validateEmail(String? value) {
     value = value.trim().toLowerCase();
     // This is just a regular expression for email addresses
     String p = "[a-zA-Z0-9\+\.\_\%\-\+]{1,256}" +
@@ -572,7 +572,7 @@ class _RegisterPageState extends State<RegisterPage> {
   }
 
 
-  String _validatePassword(String value) {
+  String _validatePassword(String? value) {
     String errorValidate = 'La contraseña es inválida, debe tener:';
     if (value.isEmpty) {
       // The form is empty
@@ -615,7 +615,7 @@ class _RegisterPageState extends State<RegisterPage> {
 
   }
 
-  String _validatePasswordConfirm(String value) {
+  String _validatePasswordConfirm(String? value) {
     if(value.isNotEmpty){
       if (_password == value){
         return null;
