@@ -2,7 +2,7 @@
 import 'package:ctpaga/views/navbar/navbar.dart';
 import 'package:ctpaga/providers/provider.dart';
 import 'package:ctpaga/env.dart';
-import 'package:flutter_masked_text/flutter_masked_text.dart';
+import 'package:flutter_masked_text2/flutter_masked_text2.dart';
 
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:provider/provider.dart';
