@@ -113,7 +113,7 @@ class _UpdatePasswordPageState extends State<UpdatePasswordPage> {
                   ),
               ),
               validator: _validatePassword,
-              onSaved: (String value) => _passwordCurrent = value,
+              onSaved: (String? value) => _passwordCurrent = value,
               textInputAction: TextInputAction.next,
               cursorColor: colorLogo,
             ),
@@ -153,7 +153,7 @@ class _UpdatePasswordPageState extends State<UpdatePasswordPage> {
                   ),
               ),
               validator: _validatePassword,
-              onSaved: (String value) => _password = value,
+              onSaved: (String? value) => _password = value,
               textInputAction: TextInputAction.next,
               cursorColor: colorLogo,
               style: TextStyle(
@@ -247,8 +247,8 @@ class _UpdatePasswordPageState extends State<UpdatePasswordPage> {
     setState(() => _statusButtonSave = true);
     await Future.delayed(Duration(milliseconds: 150));
     setState(() => _statusButtonSave = false);
-    if (_formKeyChangePassword.currentState.validate()) {
-      _formKeyChangePassword.currentState.save();
+    if (_formKeyChangePassword.currentState!.validate()) {
+      _formKeyChangePassword.currentState!.save();
       try
       {
         _onLoading();
@@ -385,7 +385,7 @@ class _UpdatePasswordPageState extends State<UpdatePasswordPage> {
   }
 
 
-  String _validatePassword(String value) {
+  String _validatePassword(String? value) {
     String errorValidate = 'La contraseña es inválida, debe tener:';
     if (value.isEmpty) {
       // The form is empty
@@ -428,7 +428,7 @@ class _UpdatePasswordPageState extends State<UpdatePasswordPage> {
 
   }
 
-  String _validatePasswordConfirm(String value) {
+  String _validatePasswordConfirm(String? value) {
     if(value.isNotEmpty){
       if (_password == value){
         return null;
