@@ -5,7 +5,7 @@ import 'package:ctpaga/views/menu/menu.dart';
 import 'package:ctpaga/providers/provider.dart';
 import 'package:ctpaga/env.dart';
 
-import 'package:flutter_masked_text/flutter_masked_text.dart';
+import 'package:flutter_masked_text2/flutter_masked_text2.dart';
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter/material.dart';
@@ -274,7 +274,7 @@ class _VerifyDataClientPageState extends State<VerifyDataClientPage> {
       else
         _total+=(priceDouble);
 
-      _totalProducts += item['quantity'];
+      _totalProducts = _totalProducts! + item['quantity'];
 
     }
 
